requirements.txt
Flask==2.2.2
matplotlib==3.5.3
numpy==1.23.3
pandas==1.4.4
Pillow==9.3.0
seaborn==0.12.1
